Sync: Fire accountRemoved KV store reason when no account (#3875)
<!--
Note: This checklist is a reminder of our shared engineering
expectations. Feel free to change it, although assigning a GitHub
reviewer and the items in bold are required.

⚠️ If you're an external contributor, please file an issue first before
working on a PR, as we can't guarantee that we will accept your changes
if they haven't been discussed ahead of time. Thanks!
-->

Task/Issue URL:
https://app.asana.com/0/1202926619870900/1209014391063056/f

**Description**:

As part of [✓ Send pixels for account removal + decoding
issues](https://app.asana.com/0/1199230911884351/1208723035104886/f),
pixels were added indicating what the reason for the Sync account being
removed from the Keychain was.

However, the pixel
`sync_account_removed_reason_not-set-on-key-value-store` is being sent
even when there was no account stored in the keychain in the first
place. This could be obscuring a problem.

To make this Pixel more useful, we should update it so that it's only
sent when there was an account stored in the keychain.

**Steps to test this PR**:
Hard to test the pixel itself as reporting on a broken state, but just
do a quick smoke test of Sync on both platforms.

<!--
Before submitting a PR, please ensure you have tested the combinations
you expect the reviewer to test, then delete configurations you *know*
do not need explicit testing.

Using a simulator where a physical device is unavailable is acceptable.
-->

**Definition of Done (Internal Only)**:

* [ ] Does this PR satisfy our [Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f)?

**Copy Testing**:

* [ ] Use of correct apostrophes in new copy, ie `’` rather than `'`

**Orientation Testing**:

* [ ] Portrait
* [ ] Landscape

**Device Testing**:

* [ ] iPhone SE (1st Gen)
* [ ] iPhone 8
* [ ] iPhone X
* [ ] iPhone 14 Pro
* [ ] iPad

**OS Testing**:

* [ ] iOS 15
* [ ] iOS 16
* [ ] iOS 17

**Theme Testing**:

* [ ] Light theme
* [ ] Dark theme

---
###### Internal references:
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)