Passwords: Update save pixel with backfilled value (#3903)
<!--
Note: This checklist is a reminder of our shared engineering
expectations. Feel free to change it, although assigning a GitHub
reviewer and the items in bold are required.

⚠️ If you're an external contributor, please file an issue first before
working on a PR, as we can't guarantee that we will accept your changes
if they haven't been discussed ahead of time. Thanks!
-->

Task/Issue URL:
https://app.asana.com/0/1208273769335188/1209241071149720/f
Tech Design URL:
CC:

**Description**:

We're appending the backfilled parameter to the _update_ pixels, but not
the _save_ pixels. So, for example, the case we're fixing for deviantart
is not being captured if you're coming in fresh. Meaning, if you don't
have credentials saved, we do backfill, but the pixel doesn't have
backfilled=true, so we can't count it.

I think at some point we started discussing only about the _update_
pixel and that was misinterpreted as don't change anything for _save_,
but that was not the intention.

The goal is to update the pixel firing logic both on macOS and iOS to
match the above requirements

**Steps to test this PR**:

1. Go to https://www.deviantart.com/users/login
2. Type in a username and hit Next
3. Type in a password and hit submit
4. You will see the Save password prompt with username and password
5. Check that you see `backfilled=true` in the save pixel on
presentation, confirmation and dismissal
6. Save the password
7. Repeat steps 1 & 2
8. Type in a new password
9. You will see the Update password prompt with username and password
10. Check that you see `backfilled=true` in the update pixel on
presentation, confirmation and dismissal

<!--
Before submitting a PR, please ensure you have tested the combinations
you expect the reviewer to test, then delete configurations you *know*
do not need explicit testing.

Using a simulator where a physical device is unavailable is acceptable.
-->

**Definition of Done (Internal Only)**:

* [x] Does this PR satisfy our [Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f)?

**Copy Testing**:

* [ ] Use of correct apostrophes in new copy, ie `’` rather than `'`

**Orientation Testing**:

* [ ] Portrait
* [ ] Landscape

**Device Testing**:

* [ ] iPhone SE (1st Gen)
* [ ] iPhone 8
* [x] iPhone X
* [ ] iPhone 14 Pro
* [ ] iPad

**OS Testing**:

* [ ] iOS 15
* [ ] iOS 16
* [x] iOS 17

**Theme Testing**:

* [ ] Light theme
* [ ] Dark theme

---
###### Internal references:
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)